+++ Procfile
@@ -1,2 +1,2 @@
 web: NODE_ENV=production node build/web
-worker: NODE_ENV=production node build/queue
+worker: NODE_ENV=production node build/worker
