+++ Procfile.dev
@@ -1,2 +1,2 @@
 web: NODE_ENV=development ./node_modules/.bin/babel-watch -L server/web
-worker: NODE_ENV=development ./node_modules/.bin/babel-watch -L server/queue
+worker: NODE_ENV=development ./node_modules/.bin/babel-watch -L server/worker
